implemented ban/unban and mute/unmute note: admins can ban/mute each others

+++ EF/Repository/UserRepository.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net.Sockets;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
@@ -22,13 +23,16 @@ namespace DATAINFRASTRUCTURE.Repository
         }
 
         public Task<User?> FindByNameAsync(string login) => _userManager.FindByNameAsync(login);
+        public Task<User?> FindByIdAsync(long userId) => _userManager.FindByIdAsync(userId.ToString());
         public Task<IdentityResult> CreateAsync(User user, string password) => _userManager.CreateAsync(user, password);
         public Task<IdentityResult> AddToRoleAsync(User user, string role) => _userManager.AddToRoleAsync(user, role);
         public Task<bool> CheckPasswordAsync(User user, string password) => _userManager.CheckPasswordAsync(user, password);
         public Task<IList<string>> GetRolesAsync(User user) => _userManager.GetRolesAsync(user);
         public Task<IdentityResult> ChangePasswordAsync(User user, string currentPassword, string newPassword)
             => _userManager.ChangePasswordAsync(user, currentPassword, newPassword);
-        public Task<IdentityResult> UpdateAsync(User user) => _userManager.UpdateAsync(user);
+        //public Task<IdentityResult> UpdateAsync(User user) => _userManager.UpdateAsync(user);
+
+
 
         //-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
         public async Task ChangeNicknameAsync(User user, string newNickname)
@@ -37,7 +41,7 @@ namespace DATAINFRASTRUCTURE.Repository
             await _userManager.UpdateAsync(user);
         }
 
-        public async Task ChangeAvatarAsync(User user, IFormFile file)
+        public async Task<bool> ChangeAvatarAsync(User user, IFormFile file)
         {
             var oldAvatarPath = user.AvatarUrl;
 
@@ -52,9 +56,51 @@ namespace DATAINFRASTRUCTURE.Repository
                 await file.CopyToAsync(stream);
 
             user.AvatarUrl = $"/avatar/{fileName}";
-            await _userManager.UpdateAsync(user);
+            var res = await _userManager.UpdateAsync(user);
 
-            File.Delete(oldAvatarPath);
+            if (res.Succeeded)
+            {
+                File.Delete(oldAvatarPath);
+                return true;
+            }
+            return false;
+        }
+
+        public async Task<bool> BanAsync(User user)
+        {
+            user.IsBanned = true;
+            var res = await _userManager.UpdateAsync(user);
+            return res.Succeeded ? true : false;
+        }
+
+        public async Task<bool> UnBanAsync(User user)
+        {
+            user.IsBanned = false;
+            var res = await _userManager.UpdateAsync(user);
+            return res.Succeeded ? true : false;
+        }
+        public async Task<bool> MuteAsync(User user)
+        {
+            user.IsMuted = true;
+            var res = await _userManager.UpdateAsync(user);
+            return res.Succeeded ? true : false;
+        }
+
+        public async Task<bool> UnMuteAsync(User user)
+        {
+            user.IsBanned = false;
+            var res = await _userManager.UpdateAsync(user);
+            return res.Succeeded ? true : false;
+        }
+
+        public async Task<bool> PromoteAsync(User user)
+        {
+            throw new NotImplementedException();
+        }
+
+        public async Task<bool> DemoteAsync(User user)
+        {
+            throw new NotImplementedException();
         }
     }
 }
